Fixed multiplier changing if grade is changed
Coursemology uses React and simply changing the value attribute of the
multiplier input will not change Coursemology's state. As such, every
time an onChange event is fired from any text input (which in
Coursemology's case, all has), all the values are re-rendered (since
changing grades and multipliers ultimately affect EXP, and they must be
displayed. This was an issue raised in React in 2017, but ultimately
deemed as a security feature. To solve this issue, we will have to
directly invoke the onChange event listener in the input prototype
(as prescribed in coursemology.js in webpack output). Read more here
https://github.com/facebook/react/issues/10135#issuecomment-314441175